scheduler/requirements.txt
#Arquivo scheduler/requirements.txt
kafka-python
requests